Extract shared normalizePath/joinPaths into tests/helpers/desktop-test-utils.js
Consolidate duplicated Node-side helper functions (normalizePath, joinPaths) from
desktop-library.spec.js and desktop-import.spec.js into a shared module. Browser-
context copies inside page.addInitScript callbacks are left inlined since they run
in Playwright's browser context and cannot use Node require().
